document EVERYTHING + small changes

+++ src/include/camera.h
@@ -9,21 +9,20 @@
 typedef struct Camera
 {
     vec3 pos;
-    vec3 dir; // direction from cam - pos + dir = target
-
-    // don't need up in cam (not useful for movement)
+    vec3 dir;   // direction from cam - pos + dir = target
+                // don't need up in cam (not useful for movement)
     vec3 right; // right perpendicularly to cam
 
     float pitch, yaw;
-    float last_cursor_x, last_cursor_y;
+    double last_cursor_x, last_cursor_y;
     float speed;
     float sensitivity;
 
     float fov;
-    float aspect_ratio; // store aspect ratio to know if it has changed
+    float aspect_ratio;
     float znear, zfar;
     mat4 view;
-    mat4 proj; // view and projection matrix
+    mat4 proj;
 } Camera;
 
 Camera camera_create(vec3 start_pos, float start_pitch, float start_yaw, float speed, float sensitivity);
